Refactored validations and other fixes
