+++ moto/glue/glue_schema_registry_constants.py
@@ -18,16 +18,18 @@ SCHEMA_VERSION_ID_PATTERN = re.compile(
 # registry constants
 MAX_REGISTRY_NAME_LENGTH = 255
 MAX_REGISTRIES_ALLOWED = 10
-CREATE_REGISTRY_OPERATION_NAME = "CreateRegistry"
 DEFAULT_REGISTRY_NAME = "default-registry"
+REGISTRY_NAME = "RegistryName"
+REGISTRY_ARN = "RegistryArn"
 
 # schema constants
 MAX_SCHEMA_NAME_LENGTH = 255
 MAX_SCHEMAS_ALLOWED = 1000
-CREATE_SCHEMA_OPERATION_NAME = "CreateSchema"
 MAX_SCHEMA_DEFINITION_LENGTH = 170000
+SCHEMA_NAME = "SchemaName"
+SCHEMA_ARN = "SchemaArn"
+SCHEMA_DEFINITION = "schemaDefinition"
 
 # schema version number constants
 MAX_VERSION_NUMBER = 100000
 MAX_SCHEMA_VERSIONS_ALLOWED = 1000
-REGISTER_SCHEMA_VERSION_OPERATION_NAME = "RegisterSchemaVersion"
